+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 0.0.*
 
+### 0.0.3
+ * Request adapter should include the correct content-type header
+ * Don't throw errors on empty responses (204s for example are valid, empty body responses)
+ * Request adapter should not send empty JSON body
+
 ### 0.0.2
  * Add `onRejected` handler for managing connectivity issues
  * Add support for custom headers at client and request level